refactor(core): collapse accepted-note repository default towers
The accepted-note write path wired repositories through five layers of
indirection: per-project factory functions, a Default dataclass wrapping
them, a build_default_* factory, and a second Default dataclass in the
mutation runner duplicating the same wiring, with a combined protocol on
top. Exactly one concrete implementation existed, and every production
call already passed repositories explicitly through
AcceptedNoteMutationDependencies.

Collapse the tower to one plain dataclass, LocalAcceptedNoteRepositories,
defined at the local composition root (deps/services.py) and constructed
once in get_note_content_mutation_service. Delete the six hidden
'repositories or build_default_accepted_note_write_repositories()'
fallbacks by making repositories a required parameter of the leaf
persistence functions, so composition is pushed up per house style. The
capability protocols (AcceptedNoteWriteRepositories,
AcceptedNoteMutationRepositories, and the per-repository protocols) stay:
they are the seams test fakes and the cloud runtime implement.

Deleted symbols (basic-memory-cloud contract):
- accepted_entity_repository_for_project
- accepted_note_content_repository_for_project
- accepted_note_search_repository_for_project
- DefaultAcceptedNoteWriteRepositories
- build_default_accepted_note_write_repositories
- AcceptedNoteRepositories (combined protocol)
- DefaultAcceptedNoteRepositories
- build_default_accepted_note_repositories
